fix: move Riccati backward pass before forward pass (#50)
This PR addresses the bug (https://github.com/TinyMPC/TinyMPC/issues/49)
that causes the solver's main loop to exit early because the terminating
criteria appears to be met even when it isn't.

This happens when setting a new `Xref` at a time when the primal and
dual residuals are already within tolerance. We need to do a
`backward_pass_grad()` first before ever checking for term. condition in
order to get updated `d` and `p`, otherwise setting the new Xref will
have no effect (= we always get the previous solution).

See issue #49.

Co-authored-by: Ishaan Mahajan <ishaanamahajan@gmail.com>